# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: FreeKick Gacha Card Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shop/buy:
    post:
      operationId: buyPack
      tags: [shop]
      summary: Purchase a pack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuyPackInput"
      responses:
        "200":
          description: Pack opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackOpenResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Insufficient coins
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /shop/nerf-preview:
    get:
      operationId: getNerfPreview
      tags: [shop]
      summary: Preview Universal Nerf Rule for a player
      parameters:
        - name: playerId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Nerf preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NerfPreview"
  /squad/{squadType}:
    get:
      operationId: getSquad
      tags: [squad]
      summary: Get a squad (main or daily_objective)
      parameters:
        - name: squadType
          in: path
          required: true
          schema:
            type: string
            enum: [main, daily_objective]
      responses:
        "200":
          description: Squad data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Squad"
  /squad/{squadType}/slot:
    put:
      operationId: setSquadSlot
      tags: [squad]
      summary: Assign a player card to a formation slot
      parameters:
        - name: squadType
          in: path
          required: true
          schema:
            type: string
            enum: [main, daily_objective]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetSlotInput"
      responses:
        "200":
          description: Updated squad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Squad"
        "422":
          description: Position incompatible
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /squad/{squadType}/slot/{slotKey}:
    delete:
      operationId: removeSquadSlot
      tags: [squad]
      summary: Remove a player from a formation slot
      parameters:
        - name: squadType
          in: path
          required: true
          schema:
            type: string
            enum: [main, daily_objective]
        - name: slotKey
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated squad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Squad"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        min_bid:
          type: integer
...
    PlayerCard:
      type: object
      required: [id, inventory_id, name, initials, position, country, club, is_gk, overall, tactical_position, is_bench_pool, is_listed, acquired_at]
      properties:
        id:
          type: integer
        inventory_id:
          type: integer
        name:
          type: string
        initials:
          type: string
        position:
          type: string
        country:
          type: string
        club:
          type: string
        is_gk:
          type: boolean
        fk:
          type: ["integer", "null"]
        pk:
          type: ["integer", "null"]
        fks:
          type: ["integer", "null"]
        pks:
          type: ["integer", "null"]
        overall:
          type: integer
        tactical_position:
          type: string
        is_bench_pool:
          type: boolean
        is_listed:
          type: boolean
        acquired_at:
          type: string
...
    BuyPackInput:
      type: object
      required: [packType]
      properties:
        packType:
          type: string
    PackOpenResult:
      type: object
      required: [players, remaining_coins, duplicate_inventory_ids]
      properties:
        players:
          type: array
          items:
            $ref: "#/components/schemas/PlayerCard"
        remaining_coins:
          type: integer
        duplicate_inventory_ids:
          type: array
          items:
            type: integer
    NerfPreview:
      type: object
      required: [player_id, name, original_role, used_as_role, effective_fk, effective_pk]
      properties:
        player_id:
          type: integer
        name:
          type: string
        original_role:
          type: string
        used_as_role:
          type: string
        effective_fk:
          type: ["number", "null"]
        effective_pk:
          type: ["number", "null"]
    SquadSlotPlayer:
      type: object
      required: [inventory_id, player_id, name, initials, overall, tactical_position, is_gk, club, country]
      properties:
        inventory_id:
          type: integer
        player_id:
          type: integer
        name:
          type: string
        initials:
          type: string
        overall:
          type: integer
        tactical_position:
          type: string
        is_gk:
          type: boolean
        fk:
          type: ["integer", "null"]
        pk:
          type: ["integer", "null"]
        fks:
          type: ["integer", "null"]
        pks:
          type: ["integer", "null"]
        club:
          type: string
        country:
          type: string
        is_bench_pool:
          type: boolean
    Squad:
      type: object
      required: [squad_type, squad_id, slots]
      properties:
        squad_type:
          type: string
        squad_id:
          type: integer
        team_ovr:
          type: ["integer", "null"]
        slots:
          type: object
          additionalProperties:
            oneOf:
              - $ref: "#/components/schemas/SquadSlotPlayer"
              - type: "null"
    SetSlotInput:
      type: object
      required: [slot_key, inventory_id]
      properties:
        slot_key:
          type: string
        inventory_id:
          type: integer
...