# lib/api-zod/src/generated/types/setSlotInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FreeKick Gacha Card Game API
 * OpenAPI spec version: 0.2.0
 */

export interface SetSlotInput {
  slot_key: string;
  inventory_id: number;
}