# lib/api-zod/src/generated/types/squad.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FreeKick Gacha Card Game API
 * OpenAPI spec version: 0.2.0
 */
import type { SquadSlots } from './squadSlots';

export interface Squad {
  squad_type: string;
  squad_id: number;
  /** @nullable */
  team_ovr?: number | null;
  slots: SquadSlots;
}